dev-requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements --dev
#

# Note: in pipenv 2020.x, "--dev" changed to emit both default and development
# requirements. To emit only development requirements, pass "--dev-only".

-i https://pypi.org/simple
asgiref==3.5.1; python_version >= '3.7'
backports.zoneinfo==0.2.1; python_version < '3.9'
black==22.1.0
brotli==1.0.9
cffi==1.15.0
click==8.1.3; python_version >= '3.7'
cryptography==37.0.2; python_version >= '3.6'
django-cors-headers==3.11.0
django==4.0.3
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
isort==5.10.1
jwt==1.3.1
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2; python_version >= '3.7'
psycopg2-binary==2.9.3
pycodestyle==2.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pycparser==2.21
pyflakes==2.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pyjwt==2.3.0; python_version >= '3.6'
python-decouple==3.6
pytz==2022.1
setuptools==62.1.0; python_version >= '3.7'
sqlparse==0.4.2; python_version >= '3.5'
tomli==2.0.1; python_version >= '3.7'
typing-extensions==4.2.0; python_version < '3.10'
whitenoise==6.0.0
